test: split URLSessionTask swizzle tests into own file
Move the URLSessionTask-specific typed swizzle tests (state and
data-task overloads) out of SentryTypedSwizzleTests into a dedicated
SentryTypedSwizzleURLSessionTasksTests, mirroring the source file split.
